+++ packages/grafana-ui/src/components/Table/BarGaugeCell.tsx
@@ -1,22 +1,14 @@
 import { isFunction } from 'lodash';
 import React from 'react';
 
-import {
-  ThresholdsConfig,
-  ThresholdsMode,
-  VizOrientation,
-  getFieldConfigWithMinMax,
-  DisplayValueAlignmentFactors,
-  Field,
-  DisplayValue,
-} from '@grafana/data';
+import { ThresholdsConfig, ThresholdsMode, VizOrientation, getFieldConfigWithMinMax } from '@grafana/data';
 import { BarGaugeDisplayMode, BarGaugeValueMode, TableCellDisplayMode } from '@grafana/schema';
 
 import { BarGauge } from '../BarGauge/BarGauge';
 import { DataLinksContextMenu, DataLinksContextMenuApi } from '../DataLinks/DataLinksContextMenu';
 
 import { TableCellProps } from './types';
-import { getCellOptions } from './utils';
+import { getAlignmentFactor, getCellOptions } from './utils';
 
 const defaultScale: ThresholdsConfig = {
   mode: ThresholdsMode.Absolute,
@@ -102,40 +94,3 @@ export const BarGaugeCell = (props: TableCellProps) => {
     </div>
   );
 };
-
-/**
- * Getting gauge values to align is very tricky without looking at all values and passing them through display processor. For very large tables that
- * could pretty expensive. So this is kind of a compromise. We look at the first 1000 rows and cache the longest value.
- * If we have a cached value we just check if the current value is longer and update the alignmentFactor. This can obviously still lead to
- * unaligned gauges but it should a lot less common.
- **/
-function getAlignmentFactor(field: Field, displayValue: DisplayValue, rowIndex: number): DisplayValueAlignmentFactors {
-  let alignmentFactor = field.state?.alignmentFactors;
-
-  if (alignmentFactor) {
-    // check if current alignmentFactor is still the longest
-    if (alignmentFactor.text.length < displayValue.text.length) {
-      alignmentFactor.text = displayValue.text;
-    }
-    return alignmentFactor;
-  } else {
-    // look at the next 1000 rows
-    alignmentFactor = { ...displayValue };
-    const maxIndex = Math.min(field.values.length, rowIndex + 1000);
-
-    for (let i = rowIndex + 1; i < maxIndex; i++) {
-      const nextDisplayValue = field.display!(field.values[i]);
-      if (nextDisplayValue.text.length > alignmentFactor.text.length) {
-        alignmentFactor.text = displayValue.text;
-      }
-    }
-
-    if (field.state) {
-      field.state.alignmentFactors = alignmentFactor;
-    } else {
-      field.state = { alignmentFactors: alignmentFactor };
-    }
-
-    return alignmentFactor;
-  }
-}
